+++ crates/outlayer/host/Cargo.toml
@@ -16,7 +16,3 @@ default = ["ed25519", "secp256k1"]
 
 ed25519 = []
 secp256k1 = []
-
-
-[dependencies]
-defuse-outlayer-types.workspace = true
