+++ crates/outlayer/host/src/ed25519.rs
@@ -1,24 +1,8 @@
-use crate::WorkerHost;
-use defuse_outlayer_types::ed25519::{Ed25519Host, Ed25519PublicKey, Ed25519Signature};
+// TODO: use defuse crypto?
+pub type Ed25519PublicKey = [u8; 32];
+pub type Ed25519Signature = [u8; 64];
 
-pub struct WorkerEd25519Host;
-
-impl Ed25519Host for WorkerEd25519Host {
-    fn ed25519_derive_public_key(&self, _path: impl AsRef<str>) -> Ed25519PublicKey {
-        unimplemented!("ed25519_derive_public_key is not implemented for WorkerHost");
-    }
-
-    fn ed25519_sign(&self, _path: impl AsRef<str>, _msg: impl AsRef<[u8]>) -> Ed25519Signature {
-        unimplemented!("ed25519_sign is not implemented for WorkerHost");
-    }
-}
-
-impl Ed25519Host for WorkerHost {
-    fn ed25519_derive_public_key(&self, path: impl AsRef<str>) -> Ed25519PublicKey {
-        self.ed25519.ed25519_derive_public_key(path)
-    }
-
-    fn ed25519_sign(&self, path: impl AsRef<str>, msg: impl AsRef<[u8]>) -> Ed25519Signature {
-        self.ed25519.ed25519_sign(path, msg)
-    }
+pub trait Ed25519Host {
+    fn ed25519_derive_public_key(&self, path: impl AsRef<str>) -> Ed25519PublicKey;
+    fn ed25519_sign(&self, path: impl AsRef<str>, msg: impl AsRef<[u8]>) -> Ed25519Signature;
 }
