+++ crates/outlayer/sdk/src/hosts/sys/ed25519.rs
