--- crates/outlayer/types/src/ed25519.rs
@@ -1,8 +1,0 @@
-// TODO: use defuse crypto?
-pub type Ed25519PublicKey = [u8; 32];
-pub type Ed25519Signature = [u8; 64];
-
-pub trait Ed25519Host {
-    fn ed25519_derive_public_key(&self, path: impl AsRef<str>) -> Ed25519PublicKey;
-    fn ed25519_sign(&self, path: impl AsRef<str>, msg: impl AsRef<[u8]>) -> Ed25519Signature;
-}
