--- crates/outlayer/types/src/lib.rs
@@ -1,4 +1,0 @@
-#[cfg(feature = "ed25519")]
-pub mod ed25519;
-#[cfg(feature = "secp256k1")]
-pub mod secp256k1;
