+++ crates/outlayer/worker-host/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "defuse-outlayer-types"
+name = "defuse-outlayer-worker-host"
 edition.workspace = true
 version.workspace = true
 rust-version.workspace = true
@@ -16,3 +16,7 @@ default = ["ed25519", "secp256k1"]
 
 ed25519 = []
 secp256k1 = []
+
+
+[dependencies]
+defuse-outlayer-host.workspace = true
